Reporting - incorporate prediction values
Fleshed out a bit more of FundProjector.  Created some functions and
properties.  Still needs to be actually implemented.

Reports now pull data from projections for date ranges beyond the
current date.

Reports now indicate the end of data. (And do not try to report for
earlier unavailble data).

Modified FundEntity to extrapolate value for the given day based on data
being in the format where we only store values of funds for the days
that CIBC releases new values.